Reword README comparison with PyOpenGL
I didn't like that the text read something like "with PyOpenGL it
takes a couple of (ie about two) lines, whereas with ModernGL it
takes several (ie more than two) lines to do the same thing."

In fixing that, I decided to split the descriptive paragraph in
two. The first sentence describes the following PyOpenGL code,
and the second sentence describes the subsequent ModernGL code.

Once that was done, the subheadings were not needed, simplifying
and condensing the appearance of the section.

Feel free to simply reject this, obviously, it's just my personal
preferences.