Addressing review comments.
1. Replace references to 'variables' with the more standard opengl
nomenclature 'attributes.'

2. Replace references to 'shaders' with 'programs', except where making
explicit reference to the vertex shader (which is commonly what recieves
the vertex attributes from our buffers).

3. The assignment of '3fi' buffer to a 'vec4' attribute is now described
in both the regular docs and in 'next', but with an explicit qualifier
that this is a ModernGL v6.0 feature. This was done so that the file
'buffer_format.rst' is identical in both the regular and the 'next'
documentation, which makes it much easier to keep them in sync.